add systemd service unit for netconfd
Allow netconfd to be started by systemd.  Assumes that persistent data
resides in /var/db/yuma.

Update configure.ac to check for systemd and only attempt to install
the service file if systemd is available.

User-specified environment variables and commandline options can be
placed in /etc/default/netconfd.

On Debian this adds a build dependency for dh_systemd which means the
packages will not build on Wheezy or earlier releases.